triangle type int

+++ Rdf/Controllers/TriangleTypeController.cs
@@ -10,7 +10,7 @@ namespace Rdf.Controllers
 {
     public class TriangleTypeController : ApiController
     {
-        public string Get(double a, double b, double c)
+        public string Get(int a, int b, int c)
         {
             var tri = new Triangle { a = a, b = b, c = c };
             return tri.GetTriangleType();
